Assign initial modules

Assigns initial modules by taking the top N target genes of each provided transcriptional regulator.

Usage

assignInitialModules(consensus_network, regulators = NULL, N = 3000L)

Arguments

consensus_network

igraph object, the consensus network across all species and replicates.

regulators

Character vector of transcriptional regulators (default: transcriptional regulators with at least 1 annotated motif in the JASPAR 2024 vertebrate core collection that are present as network nodes in the consensus network).

N

Integer, the initial module size, i.e. the number of target genes to keep for each regulator.

Value

A data frame of initial modules with 3 core columns:

regulator

Character, transcriptional regulator.

target

Character, member gene of the regulator's initial module.

weight

Numeric, consensus edge weight/adjacency, the weighted average of replicate-wise edge weights.

If the input consensus_network had more edge attributes, those appear as additional columns in the data frame.

Details

In the CroCoNet approach, the center of each network module is a transcriptional regulator. The modules are assigned in 2 main steps: 1) large initial modules are created by selecting a fixed number of target genes per regulator (performed by this function), and 2) the initial modules are pruned to keep only the best targets of each regulator (performed by pruneModules).

The module assignment is recommended to be done based on the consensus network. The regulators that provide the starting point of the module assignment can be selected based on prior biological knowledge, or the combination of prior biological knowledge and the data (see also getRegulators). As the default, the function takes all transcriptional regulators that have at least 1 annotated motif in the JASPAR 2024 vertebrate core collection (and appear in the provided consensus_network).

The function creates as many modules as there are regulators, each containing the regulator and its N best target genes. When choosing the best targets, the genes are ranked based on how strongly they are connected to the regulator (regulator-target edge weight/adjacency). N should be greater than or equal to the minimum number of targets per regulator across all regulators.

Examples

initial_modules <- assignInitialModules(consensus_network, regulators, N = 250)