Find conserved and diverged target genes within a module

Finds target genes within a conserved/diverged module that are particularly responsible for the conservation/divergence.

Usage

findConservedDivergedTargets(
  module_name,
  tree_stats_jk,
  lm_tree_stats,
  conf_level = 0.95
)

Arguments

module_name

Character, the name of the module of interest.

tree_stats_jk

Data frame of tree statistics per jackknife module version across all modules.

Columns required in case the focus of interest is conservation and overall divergence:

regulator

Character, transcriptional regulator.

module_size

Integer, the numer of target genes assigned to a regulator.

type

Character, module type (orig = original or jk = jackknifed).

id

Character, the unique ID of the module version (format: nameOfRegulator_jk_nameOfGeneRemoved in case of module type 'jk' and nameOfRegulator_orig in case of module type 'orig').

gene_removed

Character, the name of the gene removed by jackknifing (NA in case of module type 'orig').

total_tree_length

Numeric, total tree length per jackknife module version.

within_species_diversity

Numeric, within-species diversity per jackknife module version.

Columns required in case the focus of interest is species-specific divergence:

regulator

Character, transcriptional regulator.

module_size

Integer, the numer of target genes assigned to a regulator.

type

Character, module type (orig = original or jk = jackknifed).

id

Character, the unique ID of the module version (format: nameOfRegulator_jk_nameOfGeneRemoved in case of module type 'jk' and nameOfRegulator_orig in case of module type 'orig').

gene_removed

Character, the name of the gene removed by jackknifing (NA in case of module type 'orig').

{{species}}_to_other_branch_length

Numeric, the branch length between the species of interest and all others per jackknife module version

{{species}}_diversity

Numeric, diversity within the species of interest per jackknife module version.

lm_tree_stats

Object of class lm, the (weighted) linear model fit between the total tree length and within-species diversity (in case the focus of interest is conservation and overall divergence) or between the species-to-other branch length and diversity of a species (in case the focus of interest is the divergence between that particular species and all others).

conf_level

Confidence level of the prediction interval (default: 0.95).

Value

Data frame of cross-species conservation measures per target gene for the module of interest. In addition to the relevant columns of tree_stats_jk, it contains 4 new columns:

focus

Character, the focus of interest in terms of cross-species conservation, either "overall" if the focus of interest is conservation and overall divergence, or the name of a species if the focus of interest is the divergence between that particular species and all others.

fit

Numeric, the fitted total tree length/species-to-other branch length at the diversity value of a jackknife module version.

lwr_fit

Numeric, the lower bound of the prediction interval of the fit.

upr_fit

Numeric, the upper bound of the prediction interval of the fit.

residual

Numeric, the residual of the jackknife module version in the linear model. It is calculated as the difference between the observed and expected (fitted) total tree length/species-to-other branch length.

It also contains a new summary row with the aggregate values across all jackknife values of the module.

Details

To determine whether a module as whole is conserved, diverged overall or diverged on a specific lineage, the CroConet approach relies on module trees reconstrcuted from pairwise preservation scores between clones and statistics calculated based on these trees (total tree length, diversity, species-to-other branch length). To identify individual genes that contribute the most to conservation/divergence, the same statistics can be used in combination with jackknifing.

During jackknifing, each member gene of a module is removed and all statistics are recalculated, including the tree-based statistics that inform us about cross-species conservation (see the parameter jackknife in the function calculatePresStats). Our working hypothesis is that if removing a target from a diverged module makes that module more conserved, then that target was responsible for divergence in the original module, and vice versa, if removing a target from a conserved module makes that module more diverged, then that target was responsible for conservation in the original module.

To quantify these effects, the function takes use of the linear models that were fitted across all modules between the total tree length and within-species diversity (in case the focus of interest is conservation and overall divergence) or between the species-to-other branch length and the diversity within that species (in case the focus of interest is the species-specific divergence), similarly to how the conserved and diverged modules were identified in the first place. However, in this case it is not the aggregate statistic for a module as a whole that is compared to the regression line, but the statistic of each jackknife module version separately. The jackknife version that has the highest residual is the most diverged, therefore the corresponding target gene is the most conserved, while the jackknife version that has the lowest residual is the most conserved, therefore the corresponding target gene is the most diverged.

It is important to note, that in most cases the divergence/conservation of a module cannot be contributed to a single target gene, but it is rather the combined signal of all targets together. For example, it is expected that most jackknife versions of a diverged module will still fall into the diverged region above the prediction interval of the regression line, even when the most diverged target genes are removed. This also means that the "most conserved" target gene of diverged module is still most likely diverged, just to a lesser extent than the others. Therefore, it mainly makes sense to investigate the most diverged targets of the diverged modules and the most conserved targets of the conserved modules, not the other way around.

Based on the linear model provided, the function calculates the predicted value and its prediction interval at the diversity value of each jackknife module version as well as the residual of each jackknife module version. The width of the prediction interval is assumed to be constant for the entire range of diversity values across all jackknife versions of the same module and it is calculated based on the aggregate diversity value and (in case of a weighted linear model) the aggregate weight of the module as a whole (i.e. the same data that was used for fitting lm_tree_stats).

Examples

POU5F1_target_conservation <- findConservedDivergedTargets("POU5F1", tree_stats_jk, lm_overall)